sepearte out view
